userInfo 하드코딩 -> jwt userSeq 값

+++ coc/src/main/java/com/project/coc/boardHeart/mapper/BHeartMapper.java
@@ -11,5 +11,5 @@ public interface BHeartMapper {
 
     void deleteHeart(Long seq);
 
-    int heartCheck(Long boardSeq, int userInfo);
+    int heartCheck(Long boardSeq, Long userInfo);
 }
